+++ o3d3xx-o3x1xx/cpp/get_json_model/ex-get_json_model.cpp
@@ -1,8 +1,7 @@
 /*
- * Copyright 2022-present ifm electronic, gmbh
+ * Copyright 2024-present ifm electronic, gmbh
  * SPDX-License-Identifier: Apache-2.0
  */
-
 //
 // ex-get_json_model.cpp
 //