+++ o3d3xx-o3x1xx/python/o3d3xx_o3x1xx_examples/config.py
@@ -1,0 +1,12 @@
+# Configurations for the O3D3xx and 
+# O3X1xx examples.
+# Edit this file to match your setup.
+from pathlib import Path
+import os
+
+CURRENT_DIR = Path(__file__).parent.resolve().as_posix()
+
+############################################
+# Device configuration
+############################################
+IP: str = os.environ.get("IFM3D_IP", "192.168.0.22")